fix: pop in de confirmation + refactoring composants
