+++ ui/components/RDV/RdvReasons.ts
@@ -1,66 +1,52 @@
 import { EReasonsKey } from "shared"
 
-const reasons = [
+export const RdvReasons: { key: EReasonsKey; title: string }[] = [
   {
     key: EReasonsKey.MODALITE,
     title: "Modalités d'inscription",
-    checked: false,
   },
   {
     key: EReasonsKey.PLACE,
     title: "Places disponibles",
-    checked: false,
   },
   {
     key: EReasonsKey.ACCOMPAGNEMENT,
     title: "Accompagnement dans la recherche d'entreprise",
-    checked: false,
   },
   {
     key: EReasonsKey.FRAIS,
     title: "Frais d'inscription",
-    checked: false,
   },
   {
     key: EReasonsKey.CONTENU,
     title: "Contenu de la formation",
-    checked: false,
   },
   {
     key: EReasonsKey.HORAIRE,
     title: "Horaires / rythme de la formation",
-    checked: false,
   },
   {
     key: EReasonsKey.DEBOUCHE,
     title: "Situation des élèves en sortie de formation",
-    checked: false,
   },
   {
     key: EReasonsKey.PLUS,
     title: "En savoir plus sur l'alternance",
-    checked: false,
   },
   {
     key: EReasonsKey.SUIVI,
     title: "Suivi de ma candidature",
-    checked: false,
   },
   {
     key: EReasonsKey.LIEU,
     title: "Lieu de la formation",
-    checked: false,
   },
   {
     key: EReasonsKey.PORTE,
     title: "Portes ouvertes",
-    checked: false,
   },
   {
     key: EReasonsKey.AUTRE,
     title: "Autre :",
-    checked: false,
   },
 ]
-
-export { reasons }
